Use internal comms in the UKTT gem
We historically have always communicated with the backend via the
frontend application.

Ruby makes for an inefficient reverse proxy and we can actually speak
directly via the internal routes between the applications.

NB: I've disabled the CI for the moment since we no longer have public
routes available for the tests to speak to the application (the api
implementation between the backend and the frontend is slightly
different)